// types.ts
export enum AppStatus {
  ESP32_DISCONNECTED = 'ESP32_DISCONNECTED',
  ESP32_CONNECTING = 'ESP32_CONNECTING',
  MIC_PERMISSION_REQUIRED = 'MIC_PERMISSION_REQUIRED',
  READY = 'READY',
  ACTIVATING = 'ACTIVATING',
  ACTIVE_LISTENING = 'ACTIVE_LISTENING',
  ANALYZING = 'ANALYZING',
  SPEAKING = 'SPEAKING',
  ERROR = 'ERROR',
}